requirements.txt
# [Core Framework]
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1

# [AI & LangChain]
langchain-openai==0.2.0
langgraph==0.2.35
langchain-mcp-adapters==0.1.0

# [MCP Server & Client]
mcp==1.0.0
fastmcp==0.1.0

# [Tools & Utilities]
httpx==0.27.2
beautifulsoup4==4.12.3
feedparser==6.0.11
geopy==2.4.1
jinja2==3.1.4
python-multipart==0.0.12